Generalize StepGenerator and add test
