+++ cpp/src/arrow/testing/CMakeLists.txt
@@ -18,8 +18,9 @@
 arrow_install_all_headers("arrow/testing")
 
 if(ARROW_BUILD_TESTS)
-  add_arrow_test(random_test)
+  add_arrow_test(generator_test)
   add_arrow_test(gtest_util_test)
+  add_arrow_test(random_test)
 
   if(ARROW_FILESYSTEM)
     add_library(arrow_filesystem_example MODULE examplefs.cc)
